sebi_haystack/requirements.txt
haystack-ai>=2.9.0
chroma-haystack>=2.0.0
ollama-haystack>=1.1.0
chromadb>=0.5.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
streamlit>=1.40.0
pymupdf>=1.24.0
httpx>=0.27.0
pydantic>=2.0.0
